universe/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1
django-filter==2.4.0
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-taggit==1.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
markdown2==2.4.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
